+++ packages/datagrid/.storybook/preview.js
@@ -1,8 +1,6 @@
 import { namespaces as dsNamespaces } from '@talend/locales-design-system/namespaces';
 import { namespaces as componentsNamespaces } from '@talend/locales-tui-components/namespaces';
 import { namespaces as datagridNamespaces } from '@talend/locales-tui-components/namespaces';
-import cmf from '@talend/react-cmf';
-import cmfModule, { settings } from './cmfModule';
 
 export const namespaces = [dsNamespaces, componentsNamespaces, datagridNamespaces];
 
@@ -16,24 +14,9 @@ export const i18n = {
 	},
 };
 
-export const loaders = [
-	async () => {
-		const { store, saga, App } = await cmf.bootstrap({
-			render: false,
-			modules: [cmfModule],
-		});
-		saga.run();
-		store.dispatch(cmf.actions.settings.receiveSettings(settings));
-		return { store, App };
+export const parameters = {
+	chromatic: {
+		// Disable by default and enable only on some stories
+		disableSnapshot: true,
 	},
-];
-
-export const decorators = [
-	(Story, { loaded: { store, App } }) => (
-		<App store={store} registry={cmf.registry.getRegistry()}>
-			<React.Suspense fallback={null}>
-				<Story />
-			</React.Suspense>
-		</App>
-	),
-];
+};