+++ packages/datagrid/src/components/DefaultCellRenderer/AvroRenderer.test.js
@@ -3,26 +3,23 @@ import { shallow, mount } from 'enzyme';
 
 import AvroRenderer from './AvroRenderer.component';
 
+function StringDummyComponent() {
+	return <span>string dummy component</span>;
+}
+function IntDummyComponent() {
+	return <span>int dummy component</span>;
+}
+function DateDummyComponent() {
+	return <span>date dummy component</span>;
+}
+
 describe('#AvroRenderer', () => {
-	function DummyComponent() {
-		return <span>dummy component</span>;
-	}
-
-	const registryComponents = {
-		stringCellRenderer: DummyComponent,
-	};
-
-	function getComponent(id) {
-		return registryComponents[id];
-	}
-
 	it('should render AvroRenderer and load Injected Component stringCellRenderer', () => {
 		const wrapper = shallow(
 			<AvroRenderer
-				avroRenderer={{ stringCellRenderer: 'stringCellRenderer' }}
+				avroRenderer={{ stringCellRenderer: StringDummyComponent }}
 				colDef={{ avro: { type: { type: 'string' } } }}
 				data={{ value: 'value' }}
-				getComponent={getComponent}
 			/>,
 		);
 
@@ -32,133 +29,116 @@ describe('#AvroRenderer', () => {
 	it('should render AvroRenderer and load default component when no renderer', () => {
 		const wrapper = shallow(
 			<AvroRenderer
-				avroRenderer={{ stringCellRenderer: 'stringCellRenderer' }}
+				avroRenderer={{ stringCellRenderer: StringDummyComponent }}
 				colDef={{ avro: { type: { type: 'unknow' } } }}
 				data={{ value: 'value' }}
-				getComponent={getComponent}
 			/>,
 		);
 
 		expect(wrapper.getElement()).toMatchSnapshot();
 	});
 
-	it('should mount DummyComponent', () => {
+	it('should mount stringCellRenderer', () => {
 		const wrapper = mount(
 			<AvroRenderer
-				avroRenderer={{ stringCellRenderer: 'stringCellRenderer' }}
+				avroRenderer={{ stringCellRenderer: StringDummyComponent }}
 				colDef={{ avro: { type: { type: 'string' } } }}
 				data={{ value: 'value' }}
-				getComponent={getComponent}
 			/>,
 		);
 
-		expect(wrapper.find('DummyComponent').length).toBe(1);
+		expect(wrapper.find('StringDummyComponent').length).toBe(1);
 		expect(wrapper.getElement()).toMatchSnapshot();
 	});
 });
 
 describe('mapping cellRenderer', () => {
 	it('should use intCellRenderer when type is double', () => {
-		const getComponent = jest.fn();
-		shallow(
+		const wrapper = shallow(
 			<AvroRenderer
 				avroRenderer={{
-					stringCellRenderer: 'stringCellRenderer',
-					intCellRenderer: 'intCellRenderer',
+					stringCellRenderer: StringDummyComponent,
+					intCellRenderer: IntDummyComponent,
 				}}
 				colDef={{ avro: { type: { type: 'double' } } }}
 				data={{ value: 'value' }}
-				getComponent={getComponent}
 			/>,
 		);
 
-		expect(getComponent).toHaveBeenCalledWith('intCellRenderer');
+		expect(wrapper.getElement().type).toEqual(IntDummyComponent);
 	});
 
 	it('should use intCellRenderer when type is float', () => {
-		const getComponent = jest.fn();
-		shallow(
+		const wrapper = shallow(
 			<AvroRenderer
 				avroRenderer={{
-					stringCellRenderer: 'stringCellRenderer',
-					intCellRenderer: 'intCellRenderer',
+					stringCellRenderer: StringDummyComponent,
+					intCellRenderer: IntDummyComponent,
 				}}
 				colDef={{ avro: { type: { type: 'float' } } }}
 				data={{ value: 'value' }}
-				getComponent={getComponent}
 			/>,
 		);
 
-		expect(getComponent).toHaveBeenCalledWith('intCellRenderer');
+		expect(wrapper.getElement().type).toEqual(IntDummyComponent);
 	});
 
 	it('should use intCellRenderer when type is int', () => {
-		const getComponent = jest.fn();
-		shallow(
+		const wrapper = shallow(
 			<AvroRenderer
 				avroRenderer={{
-					stringCellRenderer: 'stringCellRenderer',
-					intCellRenderer: 'intCellRenderer',
+					stringCellRenderer: StringDummyComponent,
+					intCellRenderer: IntDummyComponent,
 				}}
 				colDef={{ avro: { type: { type: 'int' } } }}
 				data={{ value: 'value' }}
-				getComponent={getComponent}
 			/>,
 		);
 
-		expect(getComponent).toHaveBeenCalledWith('intCellRenderer');
+		expect(wrapper.getElement().type).toEqual(IntDummyComponent);
 	});
 
 	it('should use intCellRenderer when type is long', () => {
-		const getComponent = jest.fn();
-		shallow(
+		const wrapper = shallow(
 			<AvroRenderer
 				avroRenderer={{
-					stringCellRenderer: 'stringCellRenderer',
-					intCellRenderer: 'intCellRenderer',
+					stringCellRenderer: StringDummyComponent,
+					intCellRenderer: IntDummyComponent,
 				}}
 				colDef={{ avro: { type: { type: 'long' } } }}
 				data={{ value: 'value' }}
-				getComponent={getComponent}
 			/>,
 		);
 
-		expect(getComponent).toHaveBeenCalledWith('intCellRenderer');
+		expect(wrapper.getElement().type).toEqual(IntDummyComponent);
 	});
 
 	it('should use dateCellRenderer when type is long and logicalType is timestamp-millis', () => {
-		const getComponent = jest.fn();
-		shallow(
+		const wrapper = shallow(
 			<AvroRenderer
 				avroRenderer={{
-					stringCellRenderer: 'stringCellRenderer',
-					intCellRenderer: 'intCellRenderer',
-					dateCellRenderer: 'dateCellRenderer',
+					dateCellRenderer: DateDummyComponent,
 				}}
 				colDef={{ avro: { type: { type: 'long', logicalType: 'timestamp-millis' } } }}
 				data={{ value: 'value' }}
-				getComponent={getComponent}
 			/>,
 		);
 
-		expect(getComponent).toHaveBeenCalledWith('dateCellRenderer');
+		expect(wrapper.getElement().type).toEqual(DateDummyComponent);
 	});
 
 	it('should use intCellRenderer when type is int and logicalType is timestamp-millis', () => {
-		const getComponent = jest.fn();
-		shallow(
+		const wrapper = shallow(
 			<AvroRenderer
 				avroRenderer={{
-					stringCellRenderer: 'stringCellRenderer',
-					intCellRenderer: 'intCellRenderer',
-					dateCellRenderer: 'dateCellRenderer',
+					intCellRenderer: IntDummyComponent,
+					dateCellRenderer: DateDummyComponent,
 				}}
 				colDef={{ avro: { type: { type: 'int', logicalType: 'timestamp-millis' } } }}
 				data={{ value: 'value' }}
-				getComponent={getComponent}
 			/>,
 		);
 
-		expect(getComponent).toHaveBeenCalledWith('intCellRenderer');
+		expect(wrapper.getElement().type).toEqual(IntDummyComponent);
 	});
 });